Extracted handleANSIOutput function from replViewer.ts
